refactor: 난수 생성 기능을 구조적으로 분리 후 리팩토링
